+++ src/components/avatar/index.js
@@ -1,7 +1,5 @@
 import Avatar from './Avatar';
-import AvatarImage from './AvatarImage';
-import AvatarInitials from './AvatarInitials';
 import AvatarStack from './AvatarStack';
 
 export default Avatar;
-export { Avatar, AvatarImage, AvatarInitials, AvatarStack };
+export { Avatar, AvatarStack };